Refactor - move lib class in import dir to lib directory
