+++ lib/WebDriverProcess.js
@@ -13,8 +13,8 @@
  Must don't use protected method outside of 
  this framework.
 */
-const {ShellExecutionUtils} = require("../lib/utils/ShellExecutionUtils.js");
-const {Utils}               = require("../lib/utils/Utils.js");
+const {ShellExecutionUtils} = require("./utils/ShellExecutionUtils.js");
+const {Utils}               = require("./utils/Utils.js");
 
 class WebDriverProcess{
