+++ lib/WebElement.js
@@ -6,8 +6,8 @@
  @licence   :   GNU/GPL
  @version   :   1.0
 */
-const Exception          = require("../lib/exception/Exception.js");
-const {WebDriverRestApi} = require("../lib/restApi/WebDriverRestApi.js");
+const Exception          = require("./exception/Exception.js");
+const {WebDriverRestApi} = require("./restApi/WebDriverRestApi.js");
 const webDriverApi       = WebDriverRestApi.getInstance();
 
 class Element{
